src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Sebastian Ozimkowski <sebastianozimkowski@outlook.com>"]

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.1.5,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
drf-spectacular = ">=0.28.0,<0.29.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"